# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the kickstart for new Go APIs",
        "title": "Kickstart",
        "contact": {
            "name": "Gustavo Dias",
            "url": "https://github.com/charmingruby",
            "email": "gustavodiasa2121@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/examples": {
            "post": {
                "description": "Create a new example",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examples"
                ],
                "summary": "Create example",
                "parameters": [
                    {
                        "description": "Create Example Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.CreateExampleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/endpoint.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoint.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoint.Response"
                        }
                    }
                }
            }
        },
        "/examples/{id}": {
            "get": {
                "description": "Find an example",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examples"
                ],
                "summary": "Get example",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get Example Payload",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoint.GetExampleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/endpoint.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoint.Response"
                        }
                    }
                }
            }
        },
        "/welcome": {
            "get": {
                "description": "Health Check",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Welcome",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoint.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "endpoint.CreateExampleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "endpoint.GetExampleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/example_entity.Example"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "endpoint.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "example_entity.Example": {
            "type": "object",
            "required": [
                "created_at",
                "id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 3
                }
            }
        }
    }
}