# internal/infra/transport/rest/endpoint/get_example_endpoint.go
package endpoint

import (
	"github.com/charmingruby/kickstart/internal/core"
	"github.com/charmingruby/kickstart/internal/domain/example/example_entity"
	"github.com/gin-gonic/gin"
)

type GetExampleResponse struct {
	Message string                  `json:"message"`
	Data    *example_entity.Example `json:"data"`
	Code    int                     `json:"status_code"`
}

// GetExample godoc
//
//	@Summary		Get example
//	@Description	Find an example
//	@Tags			Examples
//	@Accept			json
//	@Produce		json
//	@Param			id	path		string	true	"Get Example Payload"
//	@Success		200	{object}	GetExampleResponse
//	@Failure		404	{object}	Response
//	@Failure		500	{object}	Response
//	@Router			/examples/{id} [get]
func (h *Handler) getExampleEndpoint(c *gin.Context) {
	exampleID := c.Param("id")

	example, err := h.exampleService.GetExample(exampleID)
	if err != nil {
		resourceNotFoundErr, ok := err.(*core.ErrNotFound)
		if ok {
			NewResourceNotFoundError(c, resourceNotFoundErr)
			return
		}

		NewInternalServerError(c, err)
		return
	}

	NewOkResponse(
		c,
		"example found",
		example,
	)
}